# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Awesome MCP Servers API",
    "version": "1.0.0",
    "description": "REST API для каталога MCP серверов",
    "contact": {
      "url": "https://mcpservers.org"
    }
  },
  "servers": [
    {
      "url": "https://mcpservers.org/api/v1",
      "description": "Production"
    }
  ],
  "paths": {
    "/servers": {
      "get": {
        "summary": "Список серверов",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 50 }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "Список серверов",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "servers": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Server" }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/servers/{id}": {
      "get": {
        "summary": "Детали сервера",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Детали сервера",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Server" }
              }
            }
          },
          "404": {
            "description": "Server not found"
          }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Поиск серверов",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Результаты поиска"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Server": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "owner": { "type": "string" },
          "repo": { "type": "string" },
          "category": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "stars": { "type": "integer" },
          "forks": { "type": "integer" },
          "isOfficial": { "type": "boolean" },
          "isRemote": { "type": "boolean" },
          "featured": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key"
      }
    }
  }
}